+++ test/test_string.rb
@@ -256,20 +256,54 @@ class CTFPartyTest < Minitest::Test
   end
 
   def test_cgi_urlencode
-    my_str = "'Stop!' said Fred"
-    assert_equal('%27Stop%21%27+said+Fred', my_str.urlencode)
+    my_url = 'http://vulnerable.site/search.aspx?txt="><script>alert(/Rubyfu/.source)</script>'
+    my_url_component = "'Stop!' said Fred"
+    assert_equal('http://vulnerable.site/search.aspx?txt=%22%3E%3Cscript%3Ealert(/Rubyfu/.source)%3C/script%3E', my_url.urlencode)
+    assert_equal("'Stop!'%20said%20Fred", my_url_component.urlencode)
   end
 
   skip def test_cgi_urlencode!
     # skip
   end
 
+  def test_cgi_urlencode_component
+    my_url = 'http://vulnerable.site/search.aspx?txt="><script>alert(/Rubyfu/.source)</script>'
+    my_url_component = "'Stop!' said Fred"
+    assert_equal("http%3A%2F%2Fvulnerable.site%2Fsearch.aspx%3Ftxt%3D%22%3E%3Cscript%3Ealert%28%2FRubyfu%2F.source%29%3C%2Fscript%3E", my_url.urlencode_component)
+    assert_equal("%27Stop%21%27+said+Fred", my_url_component.urlencode_component)
+  end
+
+  skip def test_cgi_urlencode_component!
+    # skip
+  end
+
   def test_cgi_urldecode
-    my_str = '%27Stop%21%27+said+Fred'
-    assert_equal("'Stop!' said Fred", my_str.urldecode)
+    my_url = 'http://vulnerable.site/search.aspx?txt=%22%3E%3Cscript%3Ealert(/Rubyfu/.source)%3C/script%3E'
+    my_url2 = 'http%3A%2F%2Fvulnerable.site%2Fsearch.aspx%3Ftxt%3D%22%3E%3Cscript%3Ealert%28%2FRubyfu%2F.source%29%3C%2Fscript%3E'
+    my_url_component = "'Stop!'%20said%20Fred"
+    my_url_component2 = '%27Stop%21%27+said+Fred'
+    assert_equal('http://vulnerable.site/search.aspx?txt="><script>alert(/Rubyfu/.source)</script>', my_url.urldecode)
+    assert_equal('http://vulnerable.site/search.aspx?txt="><script>alert(/Rubyfu/.source)</script>', my_url2.urldecode)
+    assert_equal("'Stop!' said Fred", my_url_component.urldecode)
+    assert_equal("'Stop!'+said+Fred", my_url_component2.urldecode)
   end
 
   skip def test_cgi_urldecode!
+    # skip
+  end
+
+  def test_cgi_urldecode_component
+    my_url = 'http://vulnerable.site/search.aspx?txt=%22%3E%3Cscript%3Ealert(/Rubyfu/.source)%3C/script%3E'
+    my_url2 = 'http%3A%2F%2Fvulnerable.site%2Fsearch.aspx%3Ftxt%3D%22%3E%3Cscript%3Ealert%28%2FRubyfu%2F.source%29%3C%2Fscript%3E'
+    my_url_component = "'Stop!'%20said%20Fred"
+    my_url_component2 = '%27Stop%21%27+said+Fred'
+    assert_equal('http://vulnerable.site/search.aspx?txt="><script>alert(/Rubyfu/.source)</script>', my_url.urldecode_component)
+    assert_equal('http://vulnerable.site/search.aspx?txt="><script>alert(/Rubyfu/.source)</script>', my_url2.urldecode_component)
+    assert_equal("'Stop!' said Fred", my_url_component.urldecode_component)
+    assert_equal("'Stop!' said Fred", my_url_component2.urldecode_component)
+  end
+
+  skip def test_cgi_urldecode_component!
     # skip
   end